Revert "Fix EnumTypedef name problem (#1091)"
This reverts commit 6ff177594d24b927408300a2ad402dceb8e9abb7.

+++ cdm/core/src/main/java/ucar/nc2/internal/iosp/hdf5/H5objects.java
@@ -1774,7 +1774,7 @@ public class H5objects {
     }
 
     int getId() {
-      return ((Short) this.properties.get(Filters.Keys.ID)).intValue();
+      return (int) this.properties.get(Filters.Keys.ID);
     }
 
     Map<String, Object> getProperties() {
